chore: move "dispenser" package to "block"
