+++ src/main/java/cn/nukkit/block/dispenser/BlockDropper.java
@@ -1,9 +1,9 @@
-package cn.nukkit.block;
+package cn.nukkit.block.dispenser;
 
+import cn.nukkit.block.BlockProperties;
+import cn.nukkit.block.BlockState;
 import cn.nukkit.blockentity.BlockEntity;
 import cn.nukkit.blockentity.BlockEntityDropper;
-import cn.nukkit.dispenser.DispenseBehavior;
-import cn.nukkit.dispenser.DropperDispenseBehavior;
 import cn.nukkit.item.Item;
 import cn.nukkit.item.ItemTool;
 import org.jetbrains.annotations.NotNull;
