+++ src/main/java/cn/nukkit/block/dispenser/DispenseBehaviorRegister.java
@@ -1,4 +1,4 @@
-package cn.nukkit.dispenser;
+package cn.nukkit.block.dispenser;
 
 import cn.nukkit.block.BlockID;
 import cn.nukkit.entity.EntityID;
