+++ src/main/java/cn/nukkit/block/dispenser/DyeDispenseBehavior.java
@@ -1,6 +1,5 @@
-package cn.nukkit.dispenser;
+package cn.nukkit.block.dispenser;
 
-import cn.nukkit.block.*;
 import cn.nukkit.block.Block;
 import cn.nukkit.item.Item;
 import cn.nukkit.math.BlockFace;
