+++ src/main/java/cn/nukkit/block/dispenser/FireworksDispenseBehavior.java
@@ -1,6 +1,5 @@
-package cn.nukkit.dispenser;
+package cn.nukkit.block.dispenser;
 
-import cn.nukkit.block.BlockDispenser;
 import cn.nukkit.entity.Entity;
 import cn.nukkit.entity.item.EntityFireworksRocket;
 import cn.nukkit.item.Item;
