+++ src/main/java/cn/nukkit/block/dispenser/ProjectileDispenseBehavior.java
@@ -1,6 +1,5 @@
-package cn.nukkit.dispenser;
+package cn.nukkit.block.dispenser;
 
-import cn.nukkit.block.BlockDispenser;
 import cn.nukkit.entity.Entity;
 import cn.nukkit.entity.projectile.EntityProjectile;
 import cn.nukkit.item.Item;
