+++ src/main/java/cn/nukkit/block/dispenser/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Classes relevant to dispenser behavior
  */
-package cn.nukkit.dispenser;
+package cn.nukkit.block.dispenser;
