+++ src/main/java/cn/nukkit/config/category/MiscSettings.java
@@ -13,7 +13,7 @@ public class MiscSettings extends OkaeriConfig {
     @Comment("pnx.settings.misc.shutdownmessage")
     String shutdownMessage = "Server closed";
     @Comment("pnx.settings.misc.installspark")
-    boolean installSpark = true;
+    boolean installSpark = false;
     @Comment("pnx.settings.misc.enableterra")
     boolean enableTerra = false;
 }
