+++ src/main/java/cn/nukkit/level/generator/object/structures/jigsaw/trialchambers/TrialChambersStructure.java
@@ -4,7 +4,7 @@ import cn.nukkit.block.Block;
 import cn.nukkit.block.BlockBarrel;
 import cn.nukkit.block.BlockChest;
 import cn.nukkit.block.BlockDecoratedPot;
-import cn.nukkit.block.BlockDispenser;
+import cn.nukkit.block.dispenser.BlockDispenser;
 import cn.nukkit.block.BlockTrialSpawner;
 import cn.nukkit.block.copper.bulb.BlockCopperBulb;
 import cn.nukkit.blockentity.BlockEntityDecoratedPot;
@@ -14,7 +14,6 @@ import cn.nukkit.entity.effect.PotionType;
 import cn.nukkit.item.Item;
 import cn.nukkit.item.ItemID;
 import cn.nukkit.level.Level;
-import cn.nukkit.level.Location;
 import cn.nukkit.level.generator.object.BlockManager;
 import cn.nukkit.level.generator.object.RandomizableContainer;
 import cn.nukkit.level.generator.object.structures.jigsaw.JigsawStructure;
