+++ src/main/java/cn/nukkit/registry/BlockRegistry.java
@@ -10,6 +10,8 @@ import cn.nukkit.block.copper.lantern.*;
 import cn.nukkit.block.copper.lightningrod.*;
 import cn.nukkit.block.customblock.CustomBlock;
 import cn.nukkit.block.customblock.CustomBlockDefinition;
+import cn.nukkit.block.dispenser.BlockDispenser;
+import cn.nukkit.block.dispenser.BlockDropper;
 import cn.nukkit.block.shelf.*;
 import cn.nukkit.education.Education;
 import cn.nukkit.item.ItemBlock;
